+++ aurora/glimpl/opengl_impl.h
@@ -87,14 +87,14 @@ namespace aurora {
 		void setTexture1DWrapProperty(ObjRefBase *pObject, TextureWrapType pWrap) override;
 		void setTexture1DWrapPropertyBorder(ObjRefBase *pObject, glm::vec3 pColor) override;
 		void setTexture1DFilter(ObjRefBase *pObject, TextureMinFilter pMin, TextureMagFilter pMag) override;
-		void updateTexture1DData(ObjRefBase *pObject, int pWidth, const float *pDataRgba) override;
+		void updateTexture1DData(ObjRefBase *pObject, int pWidth, const uint8_t *pDataRgba) override;
 		void updateTexture1DMipmap(ObjRefBase *pObject) override;
 		ObjRefBase *createTexture3D() override;
 		void destroyTexture3D(ObjRefBase *pObject) override;
 		void setTexture3DWrapProperty(ObjRefBase *pObject, TextureWrapType pWrap) override;
 		void setTexture3DWrapPropertyBorder(ObjRefBase *pObject, glm::vec3 pColor) override;
 		void setTexture3DFilter(ObjRefBase *pObject, TextureMinFilter pMin, TextureMagFilter pMag) override;
-		void updateTexture3DData(ObjRefBase *pObject, int pWidth, int pHeight, int pDepth, const float *pDataRgba) override;
+		void updateTexture3DData(ObjRefBase *pObject, int pWidth, int pHeight, int pDepth, const uint8_t *pDataRgba) override;
 		void updateTexture3DMipmap(ObjRefBase *pObject) override;
 	};
